requirements.txt
# Core data libraries
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.40

# Database connectors
psycopg2-binary==2.9.9       # PostgreSQL (binary wheel — no build tools needed)
SQLAlchemy==2.0.30            # ORM + connection pooling for PostgreSQL
pymongo==4.7.2                # MongoDB driver

# Vector database
chromadb==0.5.0               # Local persistent vector store

# LangChain ecosystem
langchain==0.2.6
langchain-openai==0.1.14      # OpenAI LLM + embeddings wrapper
langchain-community==0.2.6    # Community integrations (ChromaDB retriever etc.)
langchain-core==0.2.11

# LLM providers (only one API key needed — OpenAI or Anthropic)
openai==1.35.3

# Web / HTTP
requests==2.32.3
feedparser==6.0.11            # RSS feed parsing
beautifulsoup4==4.12.3        # HTML parsing for SEC filing text extraction

# API layer
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4

# Utilities
python-dotenv==1.0.1          # Load .env files
nbstripout==0.7.1             # Strip notebook outputs before git commits
